clean up urls
